splash: move tlonbot wait before invite screens
