+++ apps/tlon-mobile/src/fixtures/TlonbotSetupScreen.fixture.tsx
@@ -1,11 +1,10 @@
 import { FixtureWrapper } from '@tloncorp/app/fixtures/FixtureWrapper';
-
-import { TlonbotSetupScreenView } from '../screens/Onboarding/TlonbotSetupScreen';
+import { TlonBotSetupPaneView } from '@tloncorp/app/ui';
 
 export default function TlonbotSetupScreenFixture() {
   return (
     <FixtureWrapper>
-      <TlonbotSetupScreenView />
+      <TlonBotSetupPaneView />
     </FixtureWrapper>
   );
 }
