+++ packages/app/ui/index.tsx
@@ -69,6 +69,7 @@ export * from './components/Tabs';
 export * from './components/TwtitterAttestationPane';
 export * from './components/UserProfileScreenView';
 export * from './components/Wayfinding/SplashSequence';
+export * from './components/Wayfinding/TlonBotSetupPaneView';
 export * from './components/postCollectionViews/SummaryCollectionView';
 export * from './components/WidgetPane';
 export { ConnectedPostView } from './components/postCollectionViews/shared';
